backend/app/requirements.txt
aioredis==1.3.1
anyio==3.4.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
async-timeout==4.0.2
asyncpg==0.25.0
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
cryptography==36.0.1
databases==0.5.5
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
Deprecated==1.2.13
distro==1.6.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
fastapi==0.70.1
fastapi-cache==0.1.0
fastapi-redis-cache==0.2.5
h11==0.12.0
hiredis==2.0.0
httptools==0.3.0
idna==3.3
importlib-resources==5.4.0
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.0.3
json5==0.9.6
jsonschema==3.2.0
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyter-server==1.13.1
jupyterlab==3.2.5
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclassic==0.3.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.6
packaging==21.3
pandocfilters==1.5.0
paramiko==2.9.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pipreqs==0.4.11
prometheus-client==0.12.0
prompt-toolkit==3.0.24
ptyprocess==0.7.0
pycparser==2.21
pydantic==1.8.2
Pygments==2.10.0
PyNaCl==1.5.0
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==5.4.1
pyzmq==22.3.0
redis==4.1.0
requests==2.26.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
SQLAlchemy==1.4.29
starlette==0.16.0
terminado==0.12.1
testpath==0.5.0
texttable==1.6.4
tornado==6.1
traitlets==5.1.1
typing_extensions==4.0.1
urllib3==1.26.7
uvicorn==0.16.0
uvloop==0.16.0
watchgod==0.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.59.0
websockets==10.1
wrapt==1.13.3
yarg==0.1.9
zipp==3.6.0
